requirements.txt
numpy
scipy
scikit-learn==1.4.1.post1
xgboost
joblib
pandas
confluent-kafka>=2.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
